\name{celcius-package}
\alias{celcius-package}
\alias{celcius}
\docType{package}
\title{
\packageTitle{celcius}
}
\description{
\packageDescription{celcius}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{celcius}
\packageIndices{celcius}
}
\author{
\packageAuthor{celcius}

Maintainer: \packageMaintainer{celcius}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
