% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame cx with two variables v and w}

\item{W}{Weight restriction for x}
}
\value{
it returns the maximum knapsack value and which elements.
}
\description{
greedy_knapsack search is to use the a heuristic or approximation for the problem. This algorithm will not give
an exact result (but it can be shown that it will return at least 50% of the true maximum value), but it
will reduce the computational complexity considerably (actually to O(n log n) due to the sorting part of
the algorithm)
}
\examples{
knapsack_objects <- knapsack_data_gen(n = 2000, seed = 42)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
\seealso{
see also similar functions:
\code{\link{brute_force_knapsack}}
\code{\link{knapsack_dynamic}}
}

