% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_data_gen.R
\name{knapsack_data_gen}
\alias{knapsack_data_gen}
\title{Knapsack data generator}
\usage{
knapsack_data_gen(n = 2000, seed = 42)
}
\arguments{
\item{n}{number of observation to be created}

\item{seed}{The random generator seed}
}
\value{
A data frame with two columns, 'v' for value and 'w' for weight.
}
\description{
A data generator for creating knapsack elements
}

