% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data.frame cx with two variables v and w}

\item{W}{The weight capacity of the knapsack}

\item{parallel}{If true, the function will use multiple cores. Works only on Mac/Linux system.}
}
\value{
The maximum knapsack value and corresponding elements that contributes to the value.
}
\description{
A function that solve the knapsack problem through brute force. Brute-force search 
is an algorithm approach of complexity O(2n) since all possible combinations 2n needs to be evaluated.
}
\examples{
knapsack_objects <- knapsack_data_gen(n = 2000, seed = 42)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\seealso{
see also similar functions:
\code{\link{greedy_knapsack}}
\code{\link{knapsack_dynamic}}
}

