% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data.frame cx with two variables v and w}

\item{W}{Weight restriction for x}
}
\value{
it returns the maximum knapsack value and which elements.
}
\description{
Brute-force search (i.e. going through all possible alternatives and return the maximum value
found) knapsack is an algorithm approach of complexity O(2n) since all possible combinations 2n needs to be evaluated.
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\seealso{
see also similar functions:
\code{\link{greedy_knapsack}}
\code{\link{knapsack_dynamic}}
}

