\name{celcius-package}
\alias{celcius-package}
\alias{celcius}
\docType{package}
\title{
\packageTitle{celcius}
}
\description{
\packageDescription{celcius}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{celcius}
\packageIndices{celcius}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{celcius}

Maintainer: \packageMaintainer{celcius}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
