% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data.frame cx with two variables v and w}

\item{W}{Weight restriction for x}
}
\value{
it returns the maximum knapsack value and which elements.
}
\description{
If the weights are actually discrete values, knapsack_dynamic search is
 used to solve the knapsack problem but scaling much better
over all possible values of w.
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\seealso{
see also similar functions:
\code{\link{brute_force_knapsack}}
\code{\link{greedy_knapsack}}
}

